pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-poc-framework"
version = "0.1.0"
description = "PoC multi-agent orchestration skeleton for MCP operations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27",
  "pydantic>=1.10",
  "PyYAML>=6.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
